Made many changes to Crystal in order to get testing to work better.
New types:
IRunCrystalCommand - An external-to-crystal command to run some analyses on some compilation units.
ICrystalJob - A collection of crystal jobs that can be run.
ISingleCrystalJob - An internal job, to run on analysis on one compilation unit. Constructing these objects replaces runSingleUnit... or whatever it was called.
IAnalysisReporter - An interface for output from an analysis.
StandardReporter - The standard implementation of IAnalysisReporter.

New Test Classes:
AnnotatedTest - A JUnit 4 test framework, which uses the annotations given below.

New Annotations for testing:
@PassingTest - Can be used to annotate an analysis test that should pass.
@FailingTest - Can be used to annotate an analysis test that should fail, can also include the number of expected failures.
@UseAnalyses - Specifies which analyses should be used when testing a file.